moved mocks to ssh dir
